Implement a single, global memcache, removing the need for a context manager.
this ensures that caching will automatically work without the need to manually create a RezMemCache.

to accomplish this, the epoch filtering has been moved onto the various methods that require it.  this ensures that a single RezMemCache instance can be shared by multiple resolves using different timestamps within the same python session.  timestamp filtering is still done after retrieving the data from the cache, so the data placed there will remain the same.