Removed $ prefix from env var names in os.getenv
In the csh and sh plugins, the 'REZ_STORED_PROMPT', 'PS1' and 'prompt'
environment varialbe names were prefixed with '$' when passed to
os.getenv. This resulted in os.getenv looking for, for example, an
environment variable called '$REZ_STORED_PROMPT' instead of
'REZ_STORED_PROMPT', which resulted in incorrect prompts being
generated in Rez sub-shells.