-fixed bug in solver (YES) -added sortedcontainers 3rd party package (for SortedSet) -updated SOLVER.md -improved comments, varnames in solver.py
What was happening:
During the main solve loop, reductions happen against added and changed
scopes. However one case wasn't taken into account, and that's when an
existing in-conflict scope is changed into a non-conflict scope via
intersection (eg, ~foo-1 becomes foo-1). This actually needs to trigger a
reduction of foo-1 against all other scopes, because the foo scope has
WIDENED in this case. This doesn't match any other situation, where if a
scope changes, it's because it was NARROWED due to a reduction or
intersection.