refactor(rez-pip): improve the handling of bin scripts
The installation via Wheel distribution when using the --target
flag, expects the bin directory to be two directories from the base
path. After the wheel unpacking the bin directory is located in the
base directory. Move the bin directory where the RECORD file expects
it to be in order to improve readability and maintain consistency.