fix(rez-pip): ensure only the bin scripts are set as tools
The current check if a file is executable or not seems to be causing
some issues on Windows platform adding non executable files to the
tools list in the package.py. Add a further check that ensures the
executable script file is located in the bin folder.

Relates: https://github.com/nerdvegas/rez/pull/602/commits/ef14e9168077d5c31020ad4166f6adb4083a0202