Avoid hidden folder/files as it introduces problems on certain filesystems
Prefixing file/folders . on certain Linux file servers, that are mounted
via samba to Windows makes the folders hidden. But in edgecases the rename
retains the hidden attribute after the rename without the . prefix.
We observed this on rez-cp.

Some client applications (Autodesk Maya) refuses to work with hidden files.
As such avoid using hidden files for temporary artifacts for sanity sake.